+++ components/core/Services/IMaskInputConverter.cs
@@ -1,0 +1,10 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace AntDesign;
+
+public interface IMaskInputConverter
+{
+    string Convert(string value, string mask);
+}
